[FEATURE] ObjectConfiguration gets name from annotation
Currently, an object configuration's property's class must
explicitly configured with its name, even if the name is
implied by the annotation of the intended property.

This change falls back to that annotated class name, if
the name is not explicitly mentioned. For example, this
configuration now will work::

 Acme\Acme\FooService:
   properties:
     fileMonitor:
       object:
         #name: TYPO3\Flow\Monitor\FileMonitor
         arguments:
           1:
             value: 'SomeThing'

...as soon as the class FooService' property ``fileMonitor``
is appropriately annotated with ``@var \TYPO3\Flow\Monitor\FileMonitor``.

Change-Id: Ic84f2605856b43aab04672da4b75b3bc36b5e196
Resolves: #44157
Releases: master