[FEATURE] Add a way to clear caches by Path + FilePattern
Currently it's not really possible to properly implement cacheFlushing
for non flow caches because the monitoring + flushing is essentially
hardcoded in the Scripts.php + CacheManager.

This Change adds a way to easily flush a specific cache when a specific
filepath and file pattern matches.

Example:

    TYPO3_Expose_SchemaCache:
        affectedFiles:
            -
                path: 'Resources/Private/TypoScripts'
                filePattern: '\.ts2$'
            -
                path: 'Classes/'
                filePattern: '\.php$'

Now when any file inside the folder "Resources/Private/TypoScripts"
is changed in any Package it will trigger a filesChangedDetected
and the cacheManager will flush that cache if it finds a changed
file that matches the filePattern.

Releases: master
Change-Id: I4d8e110fcc152fe07ba014d64360f238ad3c0ef6