[BUGFIX] Filebackend can be racy
In some case, AJAX request per ex. the FileBackend can try to get() the
entry before it's available. It really hard to test this case, and
depend on the filesystem/IO performance.

This patch try to solve the issue by try to access the file, 10 times
with a small delay between each try.

Change-Id: Iaf0b42cf94c1e9fd4392267dda5afef4042b0be7
Related: #53262
Releases: master, 2.0