[TASK] Decouple responsibilites in default errorAction()
The default errorAction() implementation in ActionController does quite
a few things, e.g. rendering an error message and forwarding back to the
original request. When implementing an own error action, this might be
unhandy, because the default error message is not so useful, while the
whole request back-forwarding stuff works quite neat. This commit
therefore moves the different parts into their own methods, which are
called by errorAction(). Each of these calls can then be used or
omitted in custom implementations of errorAction().

Change-Id: I290e7679f72fab0426ab82810176fff5558de5f5
Resolves: #54618
Releases: master