[BUGFIX] Allow persisting of objects not tagged with entity/valueobject
The Reflection does not add properties to metadata information if
they are not tagged with entity/value object. This makes it f.e.
impossible to store an object wich implements \Serializable. The
Reflection of metadata should not reduce the information to much.

The FlowAnnotation Driver supports the type object if the object
implements \Serializable Interface ValueObject annotated Objects.
Also \DateInterval will be bind to type „time“ as doctrine uses it.

Change-Id: I6d759cffd9a46cbd789585d92324f58e66d3e9ae
Releases: master
Resolves: #54161