[BUGFIX] Prevent errors due to database change in running install
When you drop and recreate the database of a running flow
installation the migration will work fine afterwards but you
will get errors about non existing roles if you do anything
related to authorization, like creating users. This is because
the PolicyCache is not flushed and so the policies are not
imported from the Policy.yaml files again.

With this change the PolicyCache is flushed when the migration
table is created in a flow instance. This happens always when
doctrine:migrate is done on an empty database, thus preventing
the error.

Change-Id: I6689508c117dd51d8eea7a3e603e61e23866b6c5
Releases: master, 2.1, 2.0