[FEATURE] Allow creation of Resource with given identity
To allow importing of resources while keeping their identity (so that
they are not imported again and again) the ResourceTypeConverter now
allows the identity to be passed in via (__identity) when converting
from array to Resource.

This needs to be allowed via the CONFIGURATION_IDENTITY_CREATION_ALLOWED
property mapping configuration option.

Change-Id: I6693d09df7c25ac4e7ea1cf7cd565800f47df5de
Releases: master
Related: NEOS-158