[FEATURE] Allow creation of persistent objects with given identity
To allow importing of objects while keeping their identity (so that
they are not imported again and again) the PersistentObjectConverter now
allows the identity to be passed in via (__identity) when converting
from array.

This needs to be allowed via the CONFIGURATION_IDENTITY_CREATION_ALLOWED
property mapping configuration option.

Change-Id: I163b66fcb651b3035fe1f9e4dd751b929e373e14
Releases: master
Related: NEOS-158